pyproject.toml
[project]
name = "fip-ranking-parser"
version = "0.1.0"
description = "A tool to parse fip ranking data from published PDFs"
dependencies = [
    "click~=8.1.3",
    "camelot-fork~=0.20.1",
    "opencv-python~=4.7.0.72",
]
readme = "README.md"
requires-python = ">= 3.10"
license = { text = "MIT" }
classifiers = ["Private :: Do Not Upload"]

[project.scripts]
fiprank = "src.fip_ranking_parser.fiprank:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true

[tool.hatch.metadata]
allow-direct-references = true
